+++ Productivity_Timer/Productivity Timer/Activities/Models/ActivityModel.swift
@@ -7,4 +7,11 @@ struct Activity {
     var description: String?
     var isFavourite: Bool?
     var timeSpent: String?
+
+    func createNewActivity(img: String, name: String, description: String) -> Activity {
+
+        let newActivity = Activity(img: img, title: name, description: description, isFavourite: false)
+        return newActivity
+    }
+
 }
