--- Productivity_Timer/Productivity Timer/Activities/Models/NewActivityCreation.swift
@@ -1,9 +1,0 @@
-
-import UIKit
-
-func createNewActivity(img: String, name: String, description: String) -> Activity {
-
-    let newActivity = Activity(img: img, title: name, description: description, isFavourite: false)
-    return newActivity
-
-}
