--- Productivity_Timer/Productivity Timer/Activities/Models/TableCells.swift
@@ -1,5 +1,0 @@
-
-import Foundation
-
-let ActivitiesTableViewCellID = "ActivitiesTableViewCell"
-
