--- Productivity_Timer/Productivity Timer/Activities/Views/TextFields.swift
@@ -1,8 +1,0 @@
-
-import UIKit
-
-let nameTextField: UITextField = {
-    let nameTextField = UITextField()
-    nameTextField.translatesAutoresizingMaskIntoConstraints = false
-    return nameTextField
-}()
